application/flaskapp/requirements.txt
Flask
pymongo[srv]
Flask-PyMongo
python-dotenv